Constexpr constructors for `Pointer`. ToDo: constexpr Token construction from number should initialize strings with that number but not with "i". I'd like to rename/alias it to `Path`.
### History
Suddenly my program - firmware for microcontroller (uC) - becomes failing when call `operator new ()`. I thought "not enogh heap", strange idea, since there was tons of space for such a little program. So that's a time to write a cool constexpr constructors for parameters of my constant map<Pointer,StructOfFunctions> to avoid memory allocation. The chain is MyMap->RjPointer->RjPointer::Token. They are done for `Token`, but `Pointer` has non-trivial destructor and constexpr is not possible for it. The work is not so cool as I expected. But work nice. `Pointer` made of `Tokens`